rename camera nets to reflect HM01B0 function
This causes the schematic cymbol for the camera to look very strange b/c the pin mapping on the old (OV7670) camera was totally different. The advantage is that the functions appear at the proper location on the connector on the board - this helps design of future boards and software porting. (It is the closest thing to a datasheet we have for now)